all controllers working :)

+++ backend/src/controllers/outfitController.ts
@@ -1,6 +1,7 @@
 import {Request, Response} from 'express'; 
 const OutfitModel = require("../models/outfit");
 const GarmentModel = require("../models/garment");
+const UserModel = require("../models/user");
 
 const outfitCtrl = {
 
@@ -25,14 +26,29 @@ async getOutfit(req: Request, res: Response): Promise <void>  {
     }
 },
 
-async createOutfit(req: Request, res: Response): Promise<void>
+async createOutfit(req: Request, res: Response): Promise<void> 
 {
     try{
-        const newOutfit = await OutfitModel.create({ garments: [] });
+        const {id} = req.body;
+        console.log(id);
+        const user = await UserModel.findById(id).exec();
+        if(!user){
+            res.status(404).json({error: "No such user."}).send();
+            return; 
+        }
+
+        const newOutfit = await OutfitModel.create({ owner: id, garments: [] });
+        
 
         if (newOutfit == null){
             res.status(500).json({error: "Outfit creation error."}).send();
             return;
+        }
+
+        const updateRes = await UserModel.updateOne({_id: id}, {$push: {outfits: newOutfit._id}});
+        if (updateRes.modifiedCount == 0){
+            res.status(500).json({error: "Server error. (Modified error)"}).send();
+            return; 
         }
 
         res.status(200).json(newOutfit).send();
@@ -40,6 +56,7 @@ async createOutfit(req: Request, res: Response): Promise<void>
 
     }
     catch(error){
+        console.log(error);
         res.status(500).json({error: "Server error."}).send();
         return; 
     }
@@ -83,7 +100,11 @@ async deleteOutfit(req: Request, res: Response): Promise <void>  {
             res.status(404).json({error: "Outfit not found."}).send();
             return;
         }
-
+        const updateRes = await UserModel.updateOne({_id: outfit.owner}, {$pull: {outfits: id}});
+        if (updateRes.modifiedCount == 0){
+            res.status(500).json({error: "Server error. (Modified error)"}).send();
+            return; 
+        }
         const deleteRes = await OutfitModel.deleteOne({ _id: id }).exec();
         if (deleteRes.deletedCount == 0 ) { 
             throw new Error();
